requirements.txt
backtrader
pandas
finnhub-python
python-dotenv
schedule
pybind11
websocket-client